test: move autoawq conversion test to e2e
vLLM is not installed in non-e2e CI environments, so the test
under model_free/ cannot run. Moved to tests/e2e/ as a simplified
smoke test that converts the model and validates vLLM produces
coherent output.
